requirements.txt
Django==5.2.5
django-environ==0.11.2
djangorestframework==3.15.2
django-jazzmin==3.0.1
django-vite==3.1.0